requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.16.1
botocore==1.19.1
certifi==2020.6.20
cloudinary==1.22.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-passwordreset==1.1.0
django-storages==1.10.1
djangorestframework==3.12.1
gunicorn==20.0.4
isort==5.5.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==8.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
